python_code/requirement.txt
async-generator==1.10
certifi==2021.5.30
cloudpickle==1.6.0
cycler==0.10.0
Cython
dask==2021.5.0
fsspec==2021.5.0
ipython-genutils==0.2.0
locket==0.2.1
loguru==0.5.3
paho-mqtt==1.5.1
pandas==1.2.4
pandocfilters==1.4.2
partd==1.2.0
patsy==0.5.1
pybind11==2.6.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-learn
scipy
seaborn
Send2Trash==1.5.0
toolz==0.11.1
webencodings==0.5.1